Moved NavTab to the components dir I'm not sure if @jfly created NavTab before or after the components dir existed
